+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- Rename `installation` to `instance`
+
 ## [0.1.0-rc.1] - 2023-05-13
 
 - Manifest file renamed to `niterpack.toml`